modified first test case created to point to MercadoLibre
