--- src/main/java/ebay/README.md
@@ -1,4 +1,0 @@
-### Application code
- 
-These packages generally contain application code. 
-If you are writing a reusable test library, you can also place reusable test components such as Page Objects or Tasks here.
