+++ src/main/java/mercadolibre/README.md
@@ -1,0 +1,4 @@
+### Application code
+ 
+These packages generally contain application code. 
+If you are writing a reusable test library, you can also place reusable test components such as Page Objects or Tasks here.
