--- src/test/java/ebay/CucumberTestSuite.java
@@ -1,12 +1,0 @@
-package ebay;
-
-import cucumber.api.CucumberOptions;
-import net.serenitybdd.cucumber.CucumberWithSerenity;
-import org.junit.runner.RunWith;
-
-@RunWith(CucumberWithSerenity.class)
-@CucumberOptions(
-        plugin = {"pretty"},
-        features = "src/test/resources/features"
-)
-public class CucumberTestSuite {}
