--- src/test/java/ebay/utils/NavigateTo.java
@@ -1,14 +1,0 @@
-package ebay.utils;
-
-import ebay.pageobjects.EbayHomePage;
-import net.thucydides.core.annotations.Step;
-
-public class NavigateTo {
-
-    EbayHomePage ebayHomePage;
-
-    @Step("Open Ebay Homepage")
-    public void ebayHomepage() {
-        ebayHomePage.open();
-    }
-}
