+++ src/test/java/mercadolibre/CucumberTestSuite.java
@@ -1,0 +1,12 @@
+package mercadolibre;
+
+import cucumber.api.CucumberOptions;
+import net.serenitybdd.cucumber.CucumberWithSerenity;
+import org.junit.runner.RunWith;
+
+@RunWith(CucumberWithSerenity.class)
+@CucumberOptions(
+        plugin = {"pretty"},
+        features = "src/test/resources/features"
+)
+public class CucumberTestSuite {}
