+++ src/test/resources/serenity.conf
@@ -28,6 +28,6 @@ drivers {
 
 environments {
   default {
-    webdriver.base.url = "https://ebay.com"
+    webdriver.base.url = "https://www.mercadolibre.com.ar/"
   }
 }
